requirements.txt
pandas==1.5.3
playwright==1.32.1
fastapi==0.103.2
uvicorn==0.21.1
requests==2.28.2
typer==0.9.0
beautifulsoup4==4.12.2
pandas-gbq==0.19.1
Pillow==9.5.0
firebase-admin==6.1.0
google-cloud-pubsub==2.16.0
avro==1.11.1
geopy==2.3.0
google-cloud-logging==3.5.0
instructor==0.2.8
mlflow==2.7.1
plotly-express==0.4.1
google-cloud-aiplatform==1.33.1
openai==0.28.1